requirements.txt
# Web interface
streamlit>=1.24.0
pandas>=1.5.0
Pillow>=9.0.0

# API and environment
python-dotenv>=0.19.0
openai>=1.0.0

# Machine learning and similarity search
faiss-cpu>=1.7.0
numpy>=1.21.0

# Logging and utilities
loguru>=0.6.0

# Type hints support
typing-extensions>=4.0.0

# Data processing
json5>=0.9.0